+++ mvel/src/main/kotlin/org/islandoftex/arara/mvel/utils/MvelState.kt
@@ -49,25 +49,24 @@ object MvelState {
             ).forEach { name: String ->
                 map[name] = methodsKotlin.first { it.name == name }
             }
+            val ruleMethodsKotlin = RuleMethods::class.java.methods
+            listOf("getOriginalReference", "getBasename", "getSession")
+                    .forEach { name: String ->
+                        map[name] = ruleMethodsKotlin.first { it.name == name }
+                    }
         }
         map
     }
 
     /**
      * A map of directive method names to method pointers.
+     *
+     * Currently a wrapper around [conditionalMethods] but introduced for
+     * easier extensibility.
      */
     @JvmStatic
-    val directiveMethods: Map<String, Any> by lazy {
-        val map = conditionalMethods.toMutableMap()
-        kotlin.runCatching {
-            val methodsKotlin = RuleMethods::class.java.methods
-            listOf("getOriginalReference", "getBasename", "getSession")
-                    .forEach { name: String ->
-                        map[name] = methodsKotlin.first { it.name == name }
-                    }
-        }
-        map
-    }
+    val directiveMethods: Map<String, Any>
+        get() = conditionalMethods
 
     /**
      * Save the available preambles from the configuration file.